Changed the Jenkins image to derive from the official Jenkins image.
That means the Jenkins container will be based on OpenJDK. Any builds requiring Oracle JDK can be done with a Jenkins slave container with Oracle JDK installed on it. We should probably setup slave nodes anyways. Also, updated all the plugins and the version of Jenkins to the latest versions.